post merge additions

+++ include/Server.hpp
@@ -89,7 +89,7 @@ class Server {
 
  private:
   int _port;
-  int _maxBodySize;
+  size_t _maxBodySize;
   std::string _serverName;
   std::vector<Route> _routes;
   PageData _defaultErrorPage;
